pyproject.toml
[project]
name = "snippets-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=23.1.0",
    "jinja2>=3.1.5",
    "psqlpy>=0.9.0",
    "sanic[ext]>=24.12.0",
    "ua-parser>=1.0.1",
    "uvicorn>=0.34.0",
]
